+++ packages/react-querybuilder/src/types/ruleGroupsIC.ts
@@ -17,8 +17,7 @@ export type RuleGroupTypeIC<R extends RuleType = RuleType, C extends string = st
   RuleGroupType<R, C>,
   'combinator' | 'rules'
 > & {
-  // TODO: prohibit combinator property in IC groups
-  // combinator?: never;
+  combinator?: undefined;
   rules: RuleGroupICArray<RuleGroupTypeIC<R, C>, R, C>;
   /**
    * Only used when adding a rule to a query that uses independent combinators