+++ packages/react-querybuilder/src/utils/formatQuery/formatQuery.ts
@@ -17,6 +17,7 @@ import type {
 } from '../../types/index.noReact';
 import { toArray } from '../arrayUtils';
 import { convertFromIC } from '../convertQuery';
+import { isRuleGroup, isRuleGroupType } from '../isRuleGroup';
 import { isRuleOrGroupValid } from '../isRuleOrGroupValid';
 import { uniqByName } from '../uniq';
 import { defaultRuleProcessorCEL } from './defaultRuleProcessorCEL';
@@ -231,7 +232,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
         }
 
         // Groups
-        if ('rules' in rule) {
+        if (isRuleGroup(rule)) {
           return processRuleGroup(rule);
         }
 
@@ -267,7 +268,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
 
       return `${rg.not ? 'NOT ' : ''}(${processedRules
         .filter(Boolean)
-        .join('combinator' in rg ? ` ${rg.combinator} ` : ' ')})`;
+        .join(isRuleGroupType(rg) ? ` ${rg.combinator} ` : ' ')})`;
     };
 
     return processRuleGroup(ruleGroup, true);
@@ -397,7 +398,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
         if (typeof rule === 'string') {
           return rule;
         }
-        if ('rules' in rule) {
+        if (isRuleGroup(rule)) {
           return processRuleGroup(rule);
         }
         return processRule(rule);
@@ -409,7 +410,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
 
       return `${rg.not ? 'NOT ' : ''}(${processedRules
         .filter(Boolean)
-        .join('combinator' in rg ? ` ${rg.combinator} ` : ' ')})`;
+        .join(isRuleGroupType(rg) ? ` ${rg.combinator} ` : ' ')})`;
     };
 
     if (parameterized) {
@@ -432,7 +433,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
 
       const expressions: string[] = rg.rules
         .map(rule => {
-          if ('rules' in rule) {
+          if (isRuleGroup(rule)) {
             const processedRuleGroup = processRuleGroup(rule);
             if (processedRuleGroup) {
               hasChildRules = true;
@@ -462,7 +463,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
         : fallbackExpression;
     };
 
-    const rgStandard = 'combinator' in ruleGroup ? ruleGroup : convertFromIC(ruleGroup);
+    const rgStandard = isRuleGroupType(ruleGroup) ? ruleGroup : convertFromIC(ruleGroup);
     const processedQuery = processRuleGroup(rgStandard, true);
     return /^\{.+\}$/.test(processedQuery) ? processedQuery : `{${processedQuery}}`;
   }
@@ -481,7 +482,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
           if (typeof rule === 'string') {
             return celCombinatorMap[rule as DefaultCombinatorName];
           }
-          if ('rules' in rule) {
+          if (isRuleGroup(rule)) {
             return processRuleGroup(rule);
           }
           const [validationResult, fieldValidator] = validateRule(rule);
@@ -499,7 +500,9 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
         })
         .filter(Boolean)
         .join(
-          'combinator' in rg ? ` ${celCombinatorMap[rg.combinator as DefaultCombinatorName]} ` : ' '
+          isRuleGroupType(rg)
+            ? ` ${celCombinatorMap[rg.combinator as DefaultCombinatorName]} `
+            : ' '
         );
 
       const [prefix, suffix] = rg.not || !outermost ? [`${rg.not ? '!' : ''}(`, ')'] : ['', ''];
@@ -524,7 +527,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
           if (typeof rule === 'string') {
             return rule;
           }
-          if ('rules' in rule) {
+          if (isRuleGroup(rule)) {
             return processRuleGroup(rule);
           }
           const [validationResult, fieldValidator] = validateRule(rule);
@@ -541,7 +544,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
           });
         })
         .filter(Boolean)
-        .join('combinator' in rg ? ` ${rg.combinator} ` : ' ');
+        .join(isRuleGroupType(rg) ? ` ${rg.combinator} ` : ' ');
 
       const [prefix, suffix] = rg.not || !outermost ? [`${rg.not ? '!' : ''}(`, ')'] : ['', ''];
 
@@ -555,7 +558,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
    * JsonLogic
    */
   if (format === 'jsonlogic') {
-    const query = 'combinator' in ruleGroup ? ruleGroup : convertFromIC(ruleGroup);
+    const query = isRuleGroupType(ruleGroup) ? ruleGroup : convertFromIC(ruleGroup);
 
     const processRuleGroup = (rg: RuleGroupType): RQBJsonLogic => {
       if (!isRuleOrGroupValid(rg, validationMap[rg.id ?? /* istanbul ignore next */ ''])) {
@@ -564,7 +567,7 @@ function formatQuery(ruleGroup: RuleGroupTypeAny, options: FormatQueryOptions | 
 
       const processedRules = rg.rules
         .map(rule => {
-          if ('rules' in rule) {
+          if (isRuleGroup(rule)) {
             return processRuleGroup(rule);
           }
           const [validationResult, fieldValidator] = validateRule(rule);