+++ packages/react-querybuilder/src/utils/parseCEL/parseCEL.ts
@@ -10,6 +10,7 @@ import type {
   ParseCELOptions,
   ValueSource,
 } from '../../types/index.noReact';
+import { isRuleGroup } from '../isRuleGroup';
 import { fieldIsValidUtil, getFieldsArray } from '../parserUtils';
 import { celParser } from './celParser';
 import type { CELExpression, CELIdentifier, CELLiteral } from './types';
@@ -106,7 +107,7 @@ function parseCEL(cel: string, options: ParseCELOptions = {}): DefaultRuleGroupT
       if (negatedExpr) {
         if (
           !negate ||
-          (negate && !('rules' in negatedExpr) && negatedExpr.operator.startsWith('doesNot'))
+          (negate && !isRuleGroup(negatedExpr) && negatedExpr.operator.startsWith('doesNot'))
         ) {
           return ic
             ? ({ rules: [negatedExpr] } as DefaultRuleGroupTypeIC)
@@ -128,7 +129,7 @@ function parseCEL(cel: string, options: ParseCELOptions = {}): DefaultRuleGroupT
         groupOnlyIfNecessary: true,
       });
       if (rule) {
-        if ('rules' in rule || (groupOnlyIfNecessary && isCELExpressionGroup(expr.value))) {
+        if (isRuleGroup(rule) || (groupOnlyIfNecessary && isCELExpressionGroup(expr.value))) {
           return rule;
         }
         return ic ? { rules: [rule] } : { combinator: 'and', rules: [rule] };
@@ -258,7 +259,7 @@ function parseCEL(cel: string, options: ParseCELOptions = {}): DefaultRuleGroupT
   }
   const result = processCELExpression(processedCEL);
   if (result) {
-    if ('rules' in result) {
+    if (isRuleGroup(result)) {
       return result;
     }
     return { rules: [result], ...(ic ? {} : { combinator: 'and' }) };