+++ packages/react-querybuilder/src/utils/parseSQL/parseSQL.ts
@@ -8,6 +8,7 @@ import type {
   DefaultRuleType,
   ParseSQLOptions,
 } from '../../types/index.noReact';
+import { isRuleGroup } from '../isRuleGroup';
 import { fieldIsValidUtil, getFieldsArray } from '../parserUtils';
 import { sqlParser } from './sqlParser';
 import type { MixedAndXorOrList, SQLExpression, SQLIdentifier } from './types';
@@ -104,7 +105,7 @@ function parseSQL(sql: string, options: ParseSQLOptions = {}): DefaultRuleGroupT
       const rule = processSQLExpression(val);
       /* instanbul ignore else */
       if (rule) {
-        if ('rules' in rule) {
+        if (isRuleGroup(rule)) {
           return { ...rule, not: true };
         }
         return {
@@ -343,7 +344,7 @@ function parseSQL(sql: string, options: ParseSQLOptions = {}): DefaultRuleGroupT
   if (where) {
     const result = processSQLExpression(where);
     if (result) {
-      if ('rules' in result) {
+      if (isRuleGroup(result)) {
         return result;
       }
       return { rules: [result], ...(ic ? {} : { combinator: 'and' }) };